Enemy interface is now an abstract class
The enemy base being an abstract class allows me to make it conform to MonoBehaviour.

Since I am working with an abstract list of enemies that conform to this class, this should allow me to still use all that unity has to offer. The previous interface only knew what I gave it.

Idk it makes sense in my head, just go with it