redesigned death
when the enemies killed the player, they set the global death flag before waiting for the kill. this meant that every enemy would trigger a jumpscare at the same time.

Now we set a seperate value to start the jumpscare timer, which apon completing, then sets the global isDead flag to lock the player in the jumpscare